requirements.txt
azure-storage-blob==12.20.0
boto3==1.34.139
chromadb==0.5.3
datasets==2.20.0
Faker==26.0.0
Flask==3.0.3
itemadapter==0.9.0
langchain-chroma
langchain-cohere
langchain-community
langchain-core
langchain-experimental
langchain-huggingface
langchain-openai
langchain-text-splitters
langchain
langgraph
lark
matplotlib
mlflow
numpy
openai
pandas
plotly
presidio-analyzer
presidio-anonymizer
pymongo
python-dotenv
ragas
rank-bm25
reportlab
requests
seaborn
sentence-transformers
spacy
streamlit-keyup
streamlit
transformers
